Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/api/request/request-bad-port-expected.txt

Blocked access to external URL http://example.com:1/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:1/ due to access control checks.
Blocked access to external URL http://example.com:7/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:7/ due to access control checks.
Blocked access to external URL http://example.com:9/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:9/ due to access control checks.
Blocked access to external URL http://example.com:11/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:11/ due to access control checks.
Blocked access to external URL http://example.com:13/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:13/ due to access control checks.
Blocked access to external URL http://example.com:15/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:15/ due to access control checks.
Blocked access to external URL http://example.com:17/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:17/ due to access control checks.
Blocked access to external URL http://example.com:19/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:19/ due to access control checks.
Blocked access to external URL http://example.com:20/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:20/ due to access control checks.
Blocked access to external URL http://example.com:21/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:21/ due to access control checks.
Blocked access to external URL http://example.com:22/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:22/ due to access control checks.
Blocked access to external URL http://example.com:23/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:23/ due to access control checks.
Blocked access to external URL http://example.com:25/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:25/ due to access control checks.
Blocked access to external URL http://example.com:37/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:37/ due to access control checks.
Blocked access to external URL http://example.com:42/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:42/ due to access control checks.
Blocked access to external URL http://example.com:43/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:43/ due to access control checks.
Blocked access to external URL http://example.com:53/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:53/ due to access control checks.
Blocked access to external URL http://example.com:77/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:77/ due to access control checks.
Blocked access to external URL http://example.com:79/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:79/ due to access control checks.
Blocked access to external URL http://example.com:87/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:87/ due to access control checks.
Blocked access to external URL http://example.com:95/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:95/ due to access control checks.
Blocked access to external URL http://example.com:101/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:101/ due to access control checks.
Blocked access to external URL http://example.com:102/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:102/ due to access control checks.
Blocked access to external URL http://example.com:103/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:103/ due to access control checks.
Blocked access to external URL http://example.com:104/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:104/ due to access control checks.
Blocked access to external URL http://example.com:109/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:109/ due to access control checks.
Blocked access to external URL http://example.com:110/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:110/ due to access control checks.
Blocked access to external URL http://example.com:111/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:111/ due to access control checks.
Blocked access to external URL http://example.com:113/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:113/ due to access control checks.
Blocked access to external URL http://example.com:115/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:115/ due to access control checks.
Blocked access to external URL http://example.com:117/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:117/ due to access control checks.
Blocked access to external URL http://example.com:119/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:119/ due to access control checks.
Blocked access to external URL http://example.com:123/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:123/ due to access control checks.
Blocked access to external URL http://example.com:135/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:135/ due to access control checks.
Blocked access to external URL http://example.com:139/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:139/ due to access control checks.
Blocked access to external URL http://example.com:143/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:143/ due to access control checks.
Blocked access to external URL http://example.com:179/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:179/ due to access control checks.
Blocked access to external URL http://example.com:389/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:389/ due to access control checks.
Blocked access to external URL http://example.com:427/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:427/ due to access control checks.
Blocked access to external URL http://example.com:465/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:465/ due to access control checks.
Blocked access to external URL http://example.com:512/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:512/ due to access control checks.
Blocked access to external URL http://example.com:513/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:513/ due to access control checks.
Blocked access to external URL http://example.com:514/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:514/ due to access control checks.
Blocked access to external URL http://example.com:515/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:515/ due to access control checks.
Blocked access to external URL http://example.com:526/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:526/ due to access control checks.
Blocked access to external URL http://example.com:530/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:530/ due to access control checks.
Blocked access to external URL http://example.com:531/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:531/ due to access control checks.
Blocked access to external URL http://example.com:532/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:532/ due to access control checks.
Blocked access to external URL http://example.com:540/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:540/ due to access control checks.
Blocked access to external URL http://example.com:548/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:548/ due to access control checks.
Blocked access to external URL http://example.com:556/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:556/ due to access control checks.
Blocked access to external URL http://example.com:563/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:563/ due to access control checks.
Blocked access to external URL http://example.com:587/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:587/ due to access control checks.
Blocked access to external URL http://example.com:601/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:601/ due to access control checks.
Blocked access to external URL http://example.com:636/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:636/ due to access control checks.
Blocked access to external URL http://example.com:993/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:993/ due to access control checks.
Blocked access to external URL http://example.com:995/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:995/ due to access control checks.
Blocked access to external URL http://example.com:2049/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:2049/ due to access control checks.
Blocked access to external URL http://example.com:3659/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:3659/ due to access control checks.
Blocked access to external URL http://example.com:4045/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:4045/ due to access control checks.
Blocked access to external URL http://example.com:6000/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6000/ due to access control checks.
Blocked access to external URL http://example.com:6665/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6665/ due to access control checks.
Blocked access to external URL http://example.com:6666/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6666/ due to access control checks.
Blocked access to external URL http://example.com:6667/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6667/ due to access control checks.
Blocked access to external URL http://example.com:6668/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6668/ due to access control checks.
Blocked access to external URL http://example.com:6669/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6669/ due to access control checks.
Blocked access to external URL http://example.com:6697/
CONSOLE MESSAGE: line 82: Fetch API cannot load http://example.com:6697/ due to access control checks.

PASS Request on bad port 1 should throw TypeError. 
PASS Request on bad port 7 should throw TypeError. 
PASS Request on bad port 9 should throw TypeError. 
PASS Request on bad port 11 should throw TypeError. 
PASS Request on bad port 13 should throw TypeError. 
PASS Request on bad port 15 should throw TypeError. 
PASS Request on bad port 17 should throw TypeError. 
PASS Request on bad port 19 should throw TypeError. 
PASS Request on bad port 20 should throw TypeError. 
PASS Request on bad port 21 should throw TypeError. 
PASS Request on bad port 22 should throw TypeError. 
PASS Request on bad port 23 should throw TypeError. 
PASS Request on bad port 25 should throw TypeError. 
PASS Request on bad port 37 should throw TypeError. 
PASS Request on bad port 42 should throw TypeError. 
PASS Request on bad port 43 should throw TypeError. 
PASS Request on bad port 53 should throw TypeError. 
PASS Request on bad port 77 should throw TypeError. 
PASS Request on bad port 79 should throw TypeError. 
PASS Request on bad port 87 should throw TypeError. 
PASS Request on bad port 95 should throw TypeError. 
PASS Request on bad port 101 should throw TypeError. 
PASS Request on bad port 102 should throw TypeError. 
PASS Request on bad port 103 should throw TypeError. 
PASS Request on bad port 104 should throw TypeError. 
PASS Request on bad port 109 should throw TypeError. 
PASS Request on bad port 110 should throw TypeError. 
PASS Request on bad port 111 should throw TypeError. 
PASS Request on bad port 113 should throw TypeError. 
PASS Request on bad port 115 should throw TypeError. 
PASS Request on bad port 117 should throw TypeError. 
PASS Request on bad port 119 should throw TypeError. 
PASS Request on bad port 123 should throw TypeError. 
PASS Request on bad port 135 should throw TypeError. 
PASS Request on bad port 139 should throw TypeError. 
PASS Request on bad port 143 should throw TypeError. 
PASS Request on bad port 179 should throw TypeError. 
PASS Request on bad port 389 should throw TypeError. 
PASS Request on bad port 427 should throw TypeError. 
PASS Request on bad port 465 should throw TypeError. 
PASS Request on bad port 512 should throw TypeError. 
PASS Request on bad port 513 should throw TypeError. 
PASS Request on bad port 514 should throw TypeError. 
PASS Request on bad port 515 should throw TypeError. 
PASS Request on bad port 526 should throw TypeError. 
PASS Request on bad port 530 should throw TypeError. 
PASS Request on bad port 531 should throw TypeError. 
PASS Request on bad port 532 should throw TypeError. 
PASS Request on bad port 540 should throw TypeError. 
PASS Request on bad port 548 should throw TypeError. 
PASS Request on bad port 556 should throw TypeError. 
PASS Request on bad port 563 should throw TypeError. 
PASS Request on bad port 587 should throw TypeError. 
PASS Request on bad port 601 should throw TypeError. 
PASS Request on bad port 636 should throw TypeError. 
PASS Request on bad port 993 should throw TypeError. 
PASS Request on bad port 995 should throw TypeError. 
PASS Request on bad port 2049 should throw TypeError. 
PASS Request on bad port 3659 should throw TypeError. 
PASS Request on bad port 4045 should throw TypeError. 
PASS Request on bad port 6000 should throw TypeError. 
PASS Request on bad port 6665 should throw TypeError. 
PASS Request on bad port 6666 should throw TypeError. 
PASS Request on bad port 6667 should throw TypeError. 
PASS Request on bad port 6668 should throw TypeError. 
PASS Request on bad port 6669 should throw TypeError. 
PASS Request on bad port 6697 should throw TypeError. 
